scraper/requirements.txt
requests
beautifulsoup4
Flask
pdfminer.six
python-docx
